requirements.txt
Django~=3.1.14
django-autoslug==1.9.8
django-bootstrap4==2.3.1
django-ckeditor==6.0.0
django-crontab==0.7.1
django-currentuser==0.5.2
django-excel-response==2.0.5
django-filter==2.*
django-group-by==0.3.1
django-imagekit==4.1.0
django-mathfilters==1.0.0
django-mptt==0.11.0
django-online-users
django-scheduler==0.9.3
django-uuslug==1.2.0
Pillow==8.1.0
psycopg2==2.8.6
PyMySQL==1.0.2
python-magic-bin
python-slugify==4.0.1
djangorestframework~=3.14.0